refactor: split sync/src/types/mod.rs and sync/src/synchronizer/mod.rs into modular files
Split the monolithic types/mod.rs (2,011 lines) into 5 focused submodules:
- types/util.rs: TtlFilter, UnknownTxHashPriority, IBDState, post_sync_process
- types/peer.rs: HeadersSyncState, ChainSyncState, PeerFlags, HeadersSyncController, PeerState, Peers
- types/inflight_blocks.rs: TimeQuantile, TimeAnalyzer, InflightState, DownloadScheduler, InflightBlocks
- types/sync_shared.rs: SyncShared, SyncState, PendingCompactBlockMap, HeaderFieldsProvider impl
- types/active_chain.rs: ActiveChain

Split synchronizer/mod.rs (1,034 lines) into 2 focused submodules:
- synchronizer/block_fetch_cmd.rs: CanStart, FetchCMD, BlockFetchCMD
- synchronizer/synchronizer.rs: Synchronizer, CKBProtocolHandler impl, token constants

All public API paths are preserved via re-exports in mod.rs files.
No behavioral changes.
